% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{feature_emoji}
\alias{feature_emoji}
\title{Emoji features}
\usage{
feature_emoji(data, doc_id_field, text_field, top_num = 20)
}
\arguments{
\item{data}{a dataframe or tibble containing the text data and document id}

\item{doc_id_field}{unquoted field name identifying the field within the data
that represents the unique document id}

\item{text_field}{unquoted field name identifying the field name in data that
contains the text of the Tweet}

\item{top_num}{integer, the top n emojis to create features from}
}
\value{
A data frame of document ids their associated emoji features
}
\description{
Engineers features related to emojis in Tweets.
}
\details{
The full list of emojis are initially downloaded from
  https://unicode.org/Public/emoji/12.0/emoji-test.txt so an internet connection
  is required.

  The emoji features are calculated as a proportion of the total number of
  words in the Tweet:

  \itemize{
    \item{the count of the top n emojis in the data}
  }
}
\examples{
tweets <- data.frame(status_id = c(1234, 5678),
                     text = c("I tweet about one thing #onething #things",
                              "I tweet about another thing #another thing #things"),
                     stringsAsFactors = FALSE)
feature_emoji(tweets, status_id, text)
}
