% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{feature_hashtags}
\alias{feature_hashtags}
\title{Hashtag features}
\usage{
feature_hashtags(data, doc_id_field, text_field, top_n = 100L,
  min_n = 2L)
}
\arguments{
\item{data}{a dataframe or tibble containing the text data and document id}

\item{doc_id_field}{unquoted field name identifying the field within the data
that represents the unique document id}

\item{text_field}{unquoted field name identifying the field name in data that
contains the text of the Tweet}

\item{top_n}{integer, the top n hashtags to create features from}

\item{min_n}{integer, the minimum number of occurrences a hashtag must appear in the data}
}
\value{
A data frame of document id fields and their associated hashtag
  features
}
\description{
Engineers features related to hashtags in Tweets.
}
\details{
Hashtags are first converted to lowercase before the features are calculated.

The hashtag features are calculated as a proportion of the total number of words in the Tweet:
\itemize{
 \item{the count of positive/negative sentiment based on the NRC lexicon}
 \item{the count of positive/negative sentiment once all NRC lexicon sentiments are grouped}
 \item{the count of the top n hashtags used in the training data}
}
}
\examples{
tweets <- data.frame(status_id = c(1234, 5678),
                     text = c("I tweet about one thing #onething #things",
                              "I tweet about another thing #another thing #things"))
feature_hashtag(tweets, status_id, text)
}
