% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{feature_longwords}
\alias{feature_longwords}
\title{Long word features}
\usage{
feature_longwords(data, doc_id_field, text_field, top_num = 10,
  remove_urls = TRUE)
}
\arguments{
\item{data}{a dataframe or tibble containing the text data and document id}

\item{doc_id_field}{unquoted field name identifying the field within the data
that represents the unique document id}

\item{text_field}{unquoted field name identifying the field name in data that
contains the text of the Tweet}

\item{top_num}{integer, the top n long words to create features from}

\item{remove_urls}{logical; should urls be removed prior to identifying long words}
}
\value{
A data frame of document ids their associated long word features
}
\description{
Engineers features related to long words in Tweets.
}
\details{
Tweets are first converted to lowercase before the features are
  calculated. Twitter URLs (if remove_urls = TRUE), hashtags and
  @usernames are also removed.

  The long word features are calculated as a proportion of the total number of
  words in the Tweet:

  \itemize{
    \item{the count of words that have 3 or more identical consecutive characters}
    \item{the count of the top_num words that are over 10 letters long}
  }
}
\examples{
tweets <- data.frame(status_id = c(1234, 5678),
                     text = c("I tweeet about one thing #onething #things",
                              "I tweeet about otherthings #otherthings
                               #things"),
                     stringsAsFactors = FALSE)
feature_longwords(tweets, status_id, text)
}
