% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{feature_ngrams}
\alias{feature_ngrams}
\title{ngram features}
\usage{
feature_ngrams(data, doc_id_field, text_field, type = "ngrams",
  n_ngrams = 1L, top_num = 1000L)
}
\arguments{
\item{data}{a dataframe or tibble containing the text data and document id}

\item{doc_id_field}{unquoted field name identifying the field within the data
that represents the unique document id}

\item{text_field}{unquoted field name identifying the field name in data that
contains the text of the Tweet}

\item{type}{string; either ngrams of character_shingles}

\item{n_ngrams}{integer; the number in the ngram}

\item{top_num}{integer, the top n ngrams to create features from}
}
\value{
A data frame of document ids their associated ngrams features
}
\description{
Engineers features related to ngrams (both word or character) in Tweets.
}
\details{
Tweets are first converted to lowercase before the features are
  calculated.

  The ngram features are calculated as a proportion of the total number of
  ngrams in the Tweet:

  \itemize{
    \item{the number of the top_num ngrams in all Tweets}
  }
}
\examples{
tweets <- data.frame(status_id = c(1234, 5678),
                     text = c("I tweet about one thing #onething #things",
                              "I tweet about another #anotherthing
                               #things"),
                     stringsAsFactors = FALSE)
feature_ngrams(tweets, status_id, text)
}
