% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{feature_orthographic}
\alias{feature_orthographic}
\title{Orthographic features}
\usage{
feature_orthographic(data, doc_id_field, text_field)
}
\arguments{
\item{data}{a dataframe or tibble containing the text data and document id}

\item{doc_id_field}{unquoted field name identifying the field within the data
that represents the unique document id}

\item{text_field}{unquoted field name identifying the field name in data that
contains the text of the Tweet}
}
\value{
A data frame of document ids their associated orthographic features
}
\description{
Engineers orthographic features in Tweets.
}
\details{
As part of the pre-processing to calculate these features, usernames
  containing @ are removed, as are hashtags.

  The orthographic features are calculated as a proportion of the total number of
  characters in the Tweet prior to the pre-processing steps above:

  \itemize{
    \item{the count of each of the punctuation characters}
  }
}
\examples{
tweets <- data.frame(status_id = c(1234, 5678),
                     text = c("I tweet about one thing! #onething #things",
                              "I tweet about another thing!?\%$ #another thing #things"),
                     stringsAsFactors = FALSE)
feature_orthographic(tweets, status_id, text)
}
