% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{feature_PoS}
\alias{feature_PoS}
\title{Part-of-Speech features}
\usage{
feature_PoS(data, doc_id_field, text_field)
}
\arguments{
\item{data}{a dataframe or tibble containing the text data and document id}

\item{doc_id_field}{unquoted field name identifying the field within the data
that represents the unique document id}

\item{text_field}{unquoted field name identifying the field name in data that
contains the text of the Tweet}
}
\value{
A data frame of document ids their associated Part-of-Speech word features
}
\description{
Engineers features related to Part-of-Speech words in Tweets.
}
\details{
Tweets are first converted to lowercase before the features are
  calculated.

  The Part-of-Speech word features are calculated as a proportion of the total number of
  words in the Tweet:

  \itemize{
    \item{the number of words in each Part-of-Speech grouping}
  }
}
\examples{
tweets <- data.frame(status_id = c(1234, 5678),
                     text = c("I tweet about one thing #onething #things",
                              "I tweet about another #anotherthing
                               #things"),
                     stringsAsFactors = FALSE)
feature_PoS(tweets, status_id, text)
}
