% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{feature_quantitative}
\alias{feature_quantitative}
\title{Quantitative features}
\usage{
feature_quantitative(data, doc_id_field, text_field)
}
\arguments{
\item{data}{a dataframe or tibble containing the text data and document id}

\item{doc_id_field}{unquoted field name identifying the field within the data
that represents the unique document id}

\item{text_field}{unquoted field name identifying the field name in data that
contains the text of the Tweet}
}
\value{
A data frame of document ids and their associated quantitative
  features
}
\description{
Engineers quantitative features in Tweets.
}
\details{
The quantitative features are

  \itemize{
    \item{the count of characters}
    \item{the number of words}
    \item{the average word length}
    \item{the number of sentences}
    \item{the average number of words in a sentence}
  }
}
\examples{
tweets <- data.frame(status_id = c(1234, 5678),
                     text = c("I love to tweet about one thing #onething #things",
                              "I have doubts about tweeting about another thing
                              #another thing #things"),
                     stringsAsFactors = FALSE)
feature_quantitative(tweets, status_id, text)
}
