% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{feature_sentiment}
\alias{feature_sentiment}
\title{Sentiment features}
\usage{
feature_sentiment(data, doc_id_field, text_field, sentiments = c("nrc",
  "Bing-Liu", "MPQA"))
}
\arguments{
\item{data}{a dataframe or tibble containing the text data and document id}

\item{doc_id_field}{unquoted field name identifying the field within the data
that represents the unique document id}

\item{text_field}{unquoted field name identifying the field name in data that
contains the text of the Tweet}

\item{sentiments}{character, can be one or multiple of nrc, Bing-Liu or MPQA}
}
\value{
A data frame of document ids and their associated sentiment
  features
}
\description{
Engineers features related to sentiment in Tweets.
}
\details{
Tweets are first converted to lowercase before the features are
  calculated.

  The sentiment features are calculated as a proportion of the total number of
  words in the Tweet:

  \itemize{
    \item{the count of sentiment grouping by sentiment lexicon}
  }
}
\examples{
tweets <- data.frame(status_id = c(1234, 5678),
                     text = c("I love to tweet about one thing #onething #things",
                              "I have doubts about tweeting about another thing
                              #another thing #things"),
                     stringsAsFactors = FALSE)
feature_sentiment(tweets, status_id, text)
}
