% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{feature_skip_ngrams}
\alias{feature_skip_ngrams}
\title{Skip ngram features}
\usage{
feature_skip_ngrams(data, doc_id_field, text_field, n_ngrams_min = 3L,
  n_ngrams_max = 7L, n_skip_words = 4L, top_num = 1000L)
}
\arguments{
\item{data}{a dataframe or tibble containing the text data and document id}

\item{doc_id_field}{unquoted field name identifying the field within the data
that represents the unique document id}

\item{text_field}{unquoted field name identifying the field name in data that
contains the text of the Tweet}

\item{n_ngrams_min}{integer; the minimum number in the ngram}

\item{n_ngrams_max}{integer; the maximum number in the ngram}

\item{n_skip_words}{integer; the maximum number of words to skip when calculating
ngrams}

\item{top_num}{integer, the top n skip ngrams to create features from}
}
\value{
A data frame of document ids their associated skip ngrams features
}
\description{
Engineers features related to skip word ngrams in Tweets.
}
\details{
Tweets are first converted to lowercase before the features are
  calculated.

  The ngram features are calculated as a proportion of the total number of
  skip ngrams in the Tweet that contain the same number of words:

  \itemize{
    \item{the number of the top_num skip ngrams of all lengths in all Tweets}
  }
}
\examples{
tweets <- data.frame(status_id = c(1234, 5678),
                     text = c("I tweet about one thing #onething #things",
                              "I tweet about another #anotherthing
                               #things"),
                     stringsAsFactors = FALSE)
feature_skip_ngrams(tweets, status_id, text)
}
