% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_engineering.R
\name{feature_stopwords}
\alias{feature_stopwords}
\title{Stop word features}
\usage{
feature_stopwords(data, doc_id_field, text_field, top_num = 10)
}
\arguments{
\item{data}{a dataframe or tibble containing the text data and document id}

\item{doc_id_field}{unquoted field name identifying the field within the data
that represents the unique document id}

\item{text_field}{unquoted field name identifying the field name in data that
contains the text of the Tweet}

\item{top_num}{integer, the top n stop words to create features from}
}
\value{
A data frame of document ids their associated stop word features
}
\description{
Engineers features related to stop words in Tweets.
}
\details{
Tweets are first converted to lowercase before the features are
  calculated.

  The stop word features are calculated as a proportion of the total number of
  words in the Tweet:

  \itemize{
    \item{the count of the top_num stop words in all Tweets}
  }
}
\examples{
tweets <- data.frame(status_id = c(1234, 5678),
                     text = c("I tweet about one thing #onething #things",
                              "I tweet about another #anotherthing
                               #things"),
                     stringsAsFactors = FALSE)
feature_stopwords(tweets, status_id, text)
}
