% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitterfeatures.R
\docType{package}
\name{twitterfeatures}
\alias{twitterfeatures}
\alias{twitterfeatures-package}
\title{twitterfeatures: A package for feature engineering Twitter text data}
\description{
twitterfeatures: A package for feature engineering Twitter text data
}
